removed submodules to make it more idiomatic; removed folders
